refactor(Process): :recycle: Split process view in components
