+++ src/pages/Processes/Processes.enum.ts
@@ -11,8 +11,7 @@ export enum QueriesProcesses {
   GetSource = 'get-process-query',
   GetDestination = 'get-destination-process-query',
   GetProcessPairs = 'get-process-pairs',
-  GetProcessPairsTx = 'get-process-pairs-tx',
-  GetProcessPairsRx = 'get-process-pairs-rx',
+  GetProcessPairsResult = 'get-process-pairs-result',
   GetFlowPairs = 'get-process-flow-pairs',
   GetHttp = 'get-process-http-pairs',
   GetFlowPair = 'get-process-flow-pair',
