+++ src/pages/Processes/routes.tsx
@@ -4,7 +4,7 @@ import { ProcessesRoutesPaths } from './Processes.enum';
 
 const Processes = lazy(() => import(/* webpackChunkName: "processes" */ './views/Processes'));
 const Process = lazy(() => import(/* webpackChunkName: "process" */ './views/Process'));
-const ProcessFlowPairs = lazy(() => import(/* webpackChunkName: "process-flowpairs" */ './views/ProcessPairs'));
+const ProcessFlowPairs = lazy(() => import(/* webpackChunkName: "process-flowpairs" */ './views/ProcessPair'));
 
 export const processesRoutes = [
   {
